Use default fetcher
