+++ lib/api.ts
@@ -8,7 +8,6 @@ import {
   UnparsedClearSkyData,
   UnparsedForecastData,
 } from './types';
-import { Inverters } from './enode';
 import { fetcher } from './swr';
 
 /**
@@ -92,11 +91,6 @@ export const manyActualsFetcher: Fetcher<Array<ActualData>> = async (
 export const actualsFetcher: Fetcher<ActualData> = async (url: string) => {
   const unparsedData: UnparsedActualData = await fetcher(url);
   return parseActualData(unparsedData);
-};
-
-export const invertersFetcher: Fetcher<Inverters> = async (url: string) => {
-  const inverterData: Inverters = await fetcher(url);
-  return inverterData;
 };
 
 function parseClearSkyData(
