+++ lib/types.ts
@@ -101,3 +101,32 @@ export type FormPostData = {
   orientation: number;
   tilt: number;
 };
+
+export type Inverter = {
+  id: string;
+  vendor: string;
+  chargingLocationId: string | null;
+  lastSeen: string;
+  isReachable: boolean;
+  productionState: {
+    productionRate: number | null;
+    isProducing: boolean | null;
+    totalLifetimeProduction: number | null;
+    lastUpdated: string | null;
+  };
+  information: {
+    id: string;
+    brand: string;
+    model: string;
+    siteName: string;
+    installationDate: string;
+  };
+  location: {
+    longitude: number | null;
+    latitude: number | null;
+  };
+};
+
+export type Inverters = {
+  inverters: Inverter[];
+};
